+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+## [0.9.5] - 2019-08-08
+
+### Added
+
+- Added (DBMS specific) awareness of maximum table/database/column lengths into `IQuerySyntaxHelper`
+- Create / Discover methods now validate the provided names before sending Sql to the DBMS (prevents attempts to create table names that are too long for the DBMS or entities containing periods or brackets)
+
+### Fixed 
+- Oracle no longer truncates strings in GetRuntimeName to 30
+
 ## [0.9.4] - 2019-07-29
 
 ### Fixed 
@@ -42,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Drop table to work correctly with Views
 - Exists now works correctly for Views (previously it would return true if there was no view but a table with the same name)
 
-[Unreleased]: https://github.com/HicServices/FAnsiSql/compare/0.9.4...develop
+[Unreleased]: https://github.com/HicServices/FAnsiSql/compare/0.9.5...develop
+[0.9.5]: https://github.com/HicServices/FAnsiSql/compare/0.9.4...0.9.5
 [0.9.4]: https://github.com/HicServices/FAnsiSql/compare/0.9.3...0.9.4
 [0.9.3]: https://github.com/HicServices/FAnsiSql/compare/0.9.2...0.9.3
 [0.9.2]: https://github.com/HicServices/FAnsiSql/compare/v0.9.1.10...0.9.2